# Server/src/main/resources/swagger/openapi.yaml
openapi: "3.0.0"
info:
  title: webSloy
  version: "0.0.1"
servers:
  - url: http://185.154.194.124:80
tags:
  - name: Order
  - name: Catalog
  - name: User
  - name: AvailableTimes

paths:
  /catalog:
    get:
      operationId: getFullCatalog
      tags:
        - Catalog
      responses:
        '200':
          description: Success
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/Catalog"

    post:
      operationId: addItem
      tags:
        - Catalog
      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/RawItem"
      responses:
        '200':
          description: Success, return the assigned ItemId
          content:
            application/json;charset=UTF-8:
              schema:
                type: integer

  /catalog/categories:
    get:
      operationId: getCategories
      tags:
        - Catalog
      responses:
        '200':
          description: Success
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  type: string

  /catalog/categories/{category}:
    get:
      operationId: getCatalogByCategory
      tags:
        - Catalog
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/Catalog"

  /catalog/{id}:
    delete:
      operationId: deleteItemById
      tags:
        - Catalog
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
        '404':
          description: The item was not found



  /availableTimes:
    get:
      operationId: getAvailableTimes
      tags:
        - AvailableTimes
      responses:
        '200':
          description: Success
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/AvailableTimes"

    post:
      operationId: postAvailableTimes
      tags:
        - AvailableTimes
      requestBody:
        description: Time format '2024-10-10 10:00:00.000'
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/AvailableTimes"
      responses:
        '200':
          description: Success
        '400':
          description: Invalid time format

  /order:
    post:
      operationId: addOrder
      tags:
        - Order
      requestBody:
        required: true
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/RawOrder"
      responses:
        '200':
          description: Success, return the assigned orderId
          content:
            application/json;charset=UTF-8:
              schema:
                type: integer
        '404':
          description: The item id was not found



  /payment/{id}:
    get:
      operationId: getPaymentLinkByOrderId
      tags:
        - Order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json;charset=UTF-8:
              schema:
                type: string
        '404':
          description: The order was not found

  /user/{id}:
    get:
      operationId: getUserById
      tags:
        - User
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/User"
        '404':
          description: The user was not found

components:
  schemas:
    Item:
      type: object
      title: Item
      description: The position of the menu
      required:
        - itemName
        - itemId
        - itemCost
        - itemCategory
        - isAvailable
      properties:
        itemName:
          type: string
        itemId:
          type: integer
        itemCost:
          type: integer
        itemCategory:
          type: string
        isAvailable:
          type: boolean


    RawItem:
      type: object
      title: RawItem
      description: The position of the menu to add it
      required:
        - itemName
        - itemCost
        - itemCategory
        - iikoId
      properties:
        itemName:
          type: string
        itemCost:
          type: integer
        itemCategory:
          type: string
        iikoId:
          type: string


    Catalog:
      type: object
      title: Catalog
      description: Меню
      required:
        - size
        - items
      properties:
        size:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'

    RawOrder:
      type: object
      title: RawOrder
      required:
        - time
        - userId
        - items
        - orderCost
        - comment
      properties:
        time:
          type: string
        userId:
          type: integer
        items:
          type: array
          items:
            properties:
              itemId:
                type: integer
              itemNumber:
                type: integer
        orderCost:
          type: integer
        comment:
          type: string

    FullOrder:
      type: object
      title: FullOrder
      required:
        - time
        - userId
        - orderId
        - items
        - orderCost
        - comment
        - status
      properties:
        time:
          type: string
        userId:
          type: integer
        orderId:
          type: integer
        items:
          type: array
          items:
            properties:
              item:
                $ref: '#/components/schemas/Item'
              itemNumber:
                type: integer
        orderCost:
          type: integer
        comment:
          type: string
        status:
          type: string
          enum:
            - waiting for payment
            - paid and preparing
            - ready
            - received
            - cancelled

    User:
      type: object
      title: User
      description: User
      required:
        - userId
        - orders
      properties:
        user_id:
          type: integer
        orders:
          type: array
          items:
            $ref: '#/components/schemas/FullOrder'
    AvailableTimes:
      type: object
      title: AvailableTimes
      description: Available time for making an order
      required:
        - times
      properties:
        times:
          type: array
          items:
            type: string